pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "minilab"
version = "0.1.0"
description = "MiniLab: multi-agent scientific lab assistant"
authors = [{ name = "Your Name" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pyyaml>=6.0",
  "httpx>=0.27",
  "pydantic>=2.8",
  "python-dotenv>=1.0",
  "feedparser>=6.0",
  "networkx>=3.0",
  "matplotlib>=3.7",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "ruff",
  "mypy",
]

